requirements.txt
# Web Framework
streamlit>=1.29.0
streamlit-chat>=0.1.1

# API Framework
fastapi>=0.104.1
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.23

# Vector Store and Embeddings
faiss-cpu>=1.9.0
sentence-transformers>=2.2.2

# LLM Integration
openai>=1.3.7
langchain>=0.0.335
langchain-openai>=0.0.2

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
pandas>=2.1.4
numpy>=1.24.3
python-multipart>=0.0.20
